+++ src/Game/Level.cpp
@@ -150,7 +150,7 @@ void Level::render() const
 	}
 }
 
-void Level::update(const uint64_t delta)
+void Level::update(const double delta)
 {
 	for (const auto& currentMapObject : m_mapObjects)
 	{
