+++ src/Render/Sprite.h
@@ -23,14 +23,14 @@ namespace RenderEngine
 
 		struct FrameDescription
 		{
-			FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const uint64_t _duration):
+			FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const double _duration):
 				leftBottomUV(_leftBottomUV),
 				rightTopUV(_rightTopUV),
 				duration(_duration)
 			{}
 			glm::vec2 leftBottomUV;
 			glm::vec2 rightTopUV;
-			uint64_t duration;
+			double duration;
 		};
 
 		Sprite(const std::shared_ptr<Texture2D> pTexure,
@@ -46,7 +46,7 @@ namespace RenderEngine
 		void insertFrames(std::vector<FrameDescription> framesDescriptions);
 
 		void render(const glm::vec2& position,const glm::vec2& size, const float rotation, const float layer = 0.f, const size_t frameId = 0) const;
-		uint64_t getFrameDuration(const size_t frameId) const;
+		double getFrameDuration(const size_t frameId) const;
 		size_t getFrameCount() const;
 
 	protected:
