+++ src/Render/SpriteAnimator.h
@@ -10,11 +10,11 @@ namespace RenderEngine
 	public:
 		SpriteAnimator(std::shared_ptr<Sprite> pSprite);
 		size_t getCurrentFrame() const { return m_currentFrame;};
-		void update(const uint64_t delta);
+		void update(const double delta);
 	private:
 		std::shared_ptr<Sprite> m_pSprite;
 		size_t m_currentFrame;
-		uint64_t m_curentFrameDuration;
-		uint64_t m_curentAnimationTime;
+		double m_curentFrameDuration;
+		double m_curentAnimationTime;
 	};
 }
